fix locale from file incorrectly parse to luis app
