+++ packages/lu/src/utils/filehelper.ts
@@ -300,35 +300,6 @@ export function parseJSON(input: string, appType: string) {
   }
 }
 
-export function getLuisCultureFromPath(file: string): string | null {
-  let fn = path.basename(file, path.extname(file))
-  let lang = path.extname(fn).substring(1)
-  switch (lang.toLowerCase()) {
-    case 'en-us':
-    case 'ar-ar':
-    case 'zh-cn':
-    case 'nl-nl':
-    case 'fr-fr':
-    case 'fr-ca':
-    case 'de-de':
-    case 'gu-in':
-    case 'hi-in':
-    case 'it-it':
-    case 'ja-jp':
-    case 'ko-kr':
-    case 'mr-in':
-    case 'pt-br':
-    case 'es-es':
-    case 'es-mx':
-    case 'ta-in':
-    case 'te-in':
-    case 'tr-tr':
-      return lang
-    default:
-      return null
-  }
-}
-
 export function getQnACultureFromPath(file: string): string | null {
   let fn = path.basename(file, path.extname(file))
   let lang = path.extname(fn).substring(1)
