+++ packages/luis/test/fixtures/verified/interruption/dia2.lu
@@ -1,7 +1,7 @@
 
 > LUIS application information
 > !# @app.versionId = 0.1
-> !# @app.culture = en-us
+> !# @app.culture = it-it
 > !# @app.luis_schema_version = 3.2.0
