+++ packages/luis/test/fixtures/verified/interruption/main.fr-fr.lu
@@ -1,7 +1,7 @@
 
 > LUIS application information
 > !# @app.versionId = 0.1
-> !# @app.culture = en-us
+> !# @app.culture = fr-fr
 > !# @app.luis_schema_version = 3.2.0
